+++ apps/api/src/graphql/clearance/school-clearance-delete.mutation.ts
@@ -3,16 +3,34 @@ import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { SchoolClearanceDeleteResponseDto } from '../../clearance/dto/school-clearance-delete-response.dto';
 import { SchoolClearance } from '../../clearance/entity/school-clearance.entity';
 import { SchoolClearanceService } from '../../clearance/school-clearance.service';
+import {
+  type UserContext,
+  UserCtx,
+} from '../../common/auth/param-decorators/user-context.decorator';
+import { ClearancePolicyService } from '../clearance-policy.service';
 
 @Resolver(() => SchoolClearance)
 export class SchoolClearanceDeleteMutation {
-  constructor(private readonly schoolClearanceService: SchoolClearanceService) {}
+  constructor(
+    private readonly schoolClearanceService: SchoolClearanceService,
+    private readonly clearancePolicyService: ClearancePolicyService,
+  ) {}
 
   @Mutation(() => SchoolClearanceDeleteResponseDto)
   async deleteSchoolClearance(
+    @UserCtx() userContext: UserContext,
     @Args('id', { type: () => String }) id: string,
   ): Promise<SchoolClearanceDeleteResponseDto> {
-    // @todo: Validate that the clearance entry is actually available to the current user!!!
+    const clearanceEntry = await this.schoolClearanceService.findOne({ where: { id } });
+    if (!clearanceEntry) {
+      return new SchoolClearanceDeleteResponseDto(true);
+    }
+
+    this.clearancePolicyService.verifyByUser(
+      userContext,
+      clearanceEntry.idmId,
+      clearanceEntry.schoolId,
+    );
 
     await this.schoolClearanceService.delete(id);
     return new SchoolClearanceDeleteResponseDto(true);
